ENH:adds design to add user forms and add tests
