+++ admin/src/components/user/AddMultipleUsersForm.tsx
@@ -1,14 +1,11 @@
 import { ToastContainer, toast } from "react-toastify";
 import * as XLSX from "xlsx";
 import "../../css/user/addUser.css";
-interface addUserProps {
-  showModal: boolean;
-  setShowModal: Function;
-}
-const AddUser = (props: addUserProps) => {
+const AddMultipleUsersForm = () => {
   var arraylist: any[] = [];
   const handleSubmit = async () => {
     const authToken = localStorage.getItem("authToken");
+    console.log(arraylist);
     try {
       const response = await fetch(
         "https://backend-mu-plum.vercel.app/user/add",
@@ -63,7 +60,6 @@ const AddUser = (props: addUserProps) => {
         closeButton: false,
       });
     } else {
-      props.setShowModal(false);
       let fileReader = new FileReader();
       fileReader.readAsArrayBuffer(file);
       fileReader.onload = async (e) => {
@@ -83,61 +79,54 @@ const AddUser = (props: addUserProps) => {
   };
   return (
     <>
-      {props.showModal ? (
-        <>
-          <div data-testid="AddUserModel" className="AddUserModal">
-            <div className="relative w-auto my-6 mx-auto">
-              <div className="AddUserModalContainer">
-                <div className="AddUserModelUploadContainer">
-                  <h3 className="text-3xl font-semibold">Upload File</h3>
-                  <button
-                    className="AddUserModalUploadButton"
-                    onClick={() => props.setShowModal(false)}
-                  >
-                    <span className="text-black h-6 w-6 text-2xl block outline-none">
-                      x
-                    </span>
-                  </button>
-                </div>
-                <div className="relative p-6 flex-auto">
-                  <label className="AddUserModelUploadFileLabel">
-                    Upload file
-                  </label>
-                  <input
-                    type="file"
-                    id="file-upload"
-                    className="AddUserModalUploadInput"
-                    accept=".csv,application/vnd.openxmlformats-officedocument.spreadsheetml.sheet, application/vnd.ms-excel"
-                    onChange={(event) => {
-                      onChange(event);
-                    }}
-                  />
-                  <p
-                    className="mt-1 text-sm text-gray-500 dark:text-gray-300"
-                    id="file_input_help"
-                  >
-                    csv, xlsx, xls
-                  </p>
-                </div>
-                <div className="AddUserModalFooter">
-                  <button
-                    data-testid="AddUserImageClose"
-                    className="AddUserModalCloseButton bg-red-500"
-                    type="button"
-                    onClick={() => props.setShowModal(false)}
-                  >
-                    Close
-                  </button>
-                </div>
-              </div>
+      <div
+        data-testid="fileContainer"
+        className="flex h-full justify-items-center p-7 lg:p-20 md:p-20"
+      >
+        <div className="flex items-center justify-center w-full">
+          <label
+            htmlFor="dropzone-file"
+            className="flex flex-col items-center justify-center w-full h-64 border-2 border-gray-300 border-dashed rounded-lg cursor-pointer bg-gray-50 dark:hover:bg-bray-800 dark:bg-gray-700 hover:bg-gray-100 dark:border-gray-600 dark:hover:border-gray-500 dark:hover:bg-gray-600"
+          >
+            <div className="flex flex-col items-center justify-center pt-5 pb-6">
+              <svg
+                className="w-8 h-8 mb-4 text-gray-500 dark:text-gray-400"
+                aria-hidden="true"
+                xmlns="http://www.w3.org/2000/svg"
+                fill="none"
+                viewBox="0 0 20 16"
+              >
+                <path
+                  stroke="currentColor"
+                  strokeLinecap="round"
+                  strokeLinejoin="round"
+                  strokeWidth="2"
+                  d="M13 13h3a3 3 0 0 0 0-6h-.025A5.56 5.56 0 0 0 16 6.5 5.5 5.5 0 0 0 5.207 5.021C5.137 5.017 5.071 5 5 5a4 4 0 0 0 0 8h2.167M10 15V6m0 0L8 8m2-2 2 2"
+                />
+              </svg>
+              <p className="mb-2 text-lg text-gray-500 dark:text-gray-400">
+                <span className="font-semibold">Click to upload</span> or drag
+                and drop
+              </p>
+              <p className="text-xs text-gray-500 dark:text-gray-400">
+                csv, xlsx, xls
+              </p>
             </div>
-          </div>
-          <div className="opacity-25 fixed inset-0 z-40 bg-black"></div>
-        </>
-      ) : null}
+            <input
+              id="dropzone-file"
+              type="file"
+              className="hidden"
+              accept=".csv,application/vnd.openxmlformats-officedocument.spreadsheetml.sheet, application/vnd.ms-excel"
+              onChange={(event) => {
+                onChange(event);
+              }}
+            />
+          </label>
+        </div>
+      </div>
       <ToastContainer />
     </>
   );
 };
 
-export default AddUser;
+export default AddMultipleUsersForm;
